[native] Enable sql invoked function tests
